Moved the extension methods to extensions file. https://github.com/lekaha/playone-android/pull/28#pullrequestreview-100657779
