+++ mobile-ui/src/main/java/com/playone/mobile/ui/firebase/v1/Extensions.kt
@@ -1,0 +1,40 @@
+package com.playone.mobile.ui.firebase.v1
+
+import com.google.firebase.database.DataSnapshot
+import com.google.firebase.database.DatabaseError
+import com.google.firebase.database.DatabaseReference
+import com.google.firebase.database.MutableData
+import com.google.firebase.database.Query
+import com.google.firebase.database.Transaction
+import com.playone.mobile.remote.OperationResultCallback
+import com.playone.mobile.ui.firebase.ext.addListenerForSingleValueEvent
+
+internal fun DatabaseError.makeCallback(errorCallback: FirebaseErrorCallback) =
+    errorCallback(code, message, details)
+
+internal fun <D> Query.addStrategyListener(
+    callback: PlayoneCallback<D>,
+    errorCallback: FirebaseErrorCallback,
+    strategy: DataSnapStrategy<D>
+) = addListenerForSingleValueEvent {
+    onDataChange = { strategy?.apply { callback(this(it)) } }
+    onCancelled = { it.makeCallback(errorCallback) }
+}
+
+internal fun <D> DatabaseReference.runTransaction(
+    callback: OperationResultCallback,
+    errorCallback: FirebaseErrorCallback,
+    strategy: TransactionDataSnapStrategy<D>,
+    block: (MutableData?) -> Transaction.Result
+) = runTransaction(object : Transaction.Handler {
+    override fun onComplete(de: DatabaseError, p1: Boolean, ds: DataSnapshot?) =
+        if (!p1) {
+            de.makeCallback(errorCallback)
+        }
+        else {
+            strategy?.invoke(de, p1, ds)
+            callback(p1)
+        }
+
+    override fun doTransaction(mutableData: MutableData?) = block(mutableData)
+})
